+++ crates/radicle-cli/src/commands/config/args.rs
@@ -29,6 +29,8 @@ pub(crate) enum Command {
     /// Open the config in your editor
     Edit,
     /// Get a value from the current configuration
+    // Command is obsolete.
+    #[command(hide = true)]
     Get {
         /// The JSON key path to the value you want to get
         key: String,
@@ -36,6 +38,8 @@ pub(crate) enum Command {
     /// Prints the JSON Schema of the Radicle configuration
     Schema,
     /// Set a key to a value in the current configuration
+    // Command is obsolete.
+    #[command(hide = true)]
     Set {
         /// The JSON key path to the value you want to set
         key: String,
@@ -43,12 +47,16 @@ pub(crate) enum Command {
         value: String,
     },
     /// Set a key in the current configuration to `null`
+    // Command is obsolete.
+    #[command(hide = true)]
     Unset {
         /// The JSON key path to the value you want to unset
         key: String,
     },
     /// Push a value onto an array, which is identified by the key, in the
     /// current configuration
+    // Command is obsolete.
+    #[command(hide = true)]
     Push {
         /// The JSON key path to the array you want to push to
         key: String,
@@ -59,6 +67,8 @@ pub(crate) enum Command {
     /// current configuration
     ///
     /// All instances of the value in the array will be removed
+    // Command is obsolete.
+    #[command(hide = true)]
     Remove {
         /// The JSON key path to the array you want to push to
         key: String,